% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{combine_predictions}
\alias{combine_predictions}
\title{Combine Prediction Results}
\usage{
combine_predictions(predictions)
}
\arguments{
\item{predictions}{list of prediction results}
}
\description{
Combine Prediction Results
}
\keyword{internal}
