% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{create_dummy_rows}
\alias{create_dummy_rows}
\title{Create Dummy Rows for Ordinal Variables}
\usage{
create_dummy_rows(data, ov_ord)
}
\arguments{
\item{data}{data.frame}

\item{ov_ord}{ordinal variable names}
}
\value{
data.frame with dummy rows
}
\description{
Create Dummy Rows for Ordinal Variables
}
\keyword{internal}
