% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_single_group.R
\name{is_single_group}
\alias{is_single_group}
\title{Check if a lavaan model is single-group}
\usage{
is_single_group(fit)
}
\arguments{
\item{fit}{An object of class \code{lavaan}.}
}
\value{
Logical scalar. Returns \code{TRUE} if the model has only
one group, \code{FALSE} otherwise.
}
\description{
This helper function checks whether a fitted \code{lavaan} model
contains only a single group.
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3 '
fit <- lavaan::cfa(HS.model, data = lavaan::HolzingerSwineford1939)
is_single_group(fit)

}
