% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{process_chunks_parallel}
\alias{process_chunks_parallel}
\title{Process Chunks in Parallel}
\usage{
process_chunks_parallel(object, data, dummy_data, idx_list, group_id, ...)
}
\arguments{
\item{object}{lavaan object}

\item{data}{original data}

\item{dummy_data}{dummy rows}

\item{idx_list}{list of row indices for chunks}

\item{group_id}{group identifier (NULL for combined processing)}

\item{...}{additional arguments}
}
\description{
Process Chunks in Parallel
}
\keyword{internal}
