% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{process_combined}
\alias{process_combined}
\title{Process Combined (All Groups Together)}
\usage{
process_combined(object, dat, ov_ord, chunk_size, ...)
}
\arguments{
\item{object}{lavaan object}

\item{dat}{data}

\item{ov_ord}{ordinal variable names}

\item{chunk_size}{chunk size}

\item{...}{additional arguments}
}
\description{
Process Combined (All Groups Together)
}
\keyword{internal}
