% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{process_multigroup_by_group}
\alias{process_multigroup_by_group}
\title{Process Multi-group Model by Group}
\usage{
process_multigroup_by_group(object, dat, ov_ord, chunk_size, ...)
}
\arguments{
\item{object}{lavaan object}

\item{dat}{data}

\item{ov_ord}{ordinal variable names}

\item{chunk_size}{chunk size}

\item{...}{additional arguments}
}
\description{
Process Multi-group Model by Group
}
\keyword{internal}
