% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_lavPredict.R
\name{process_single_group}
\alias{process_single_group}
\title{Process Single Group}
\usage{
process_single_group(object, group_data, ov_ord, chunk_size, group_id, ...)
}
\arguments{
\item{object}{lavaan object}

\item{group_data}{data for this group}

\item{ov_ord}{ordinal variable names}

\item{chunk_size}{chunk size}

\item{group_id}{group identifier}

\item{...}{additional arguments}
}
\description{
Process Single Group
}
\keyword{internal}
