% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{model_info}
\alias{model_info}
\title{Summarize basic model properties from a lavaan/blavaan fit}
\usage{
model_info(fit)
}
\arguments{
\item{fit}{A fitted lavaan or blavaan object.}
}
\value{
A named list with model properties (groups, categorical, multilevel, etc.).
}
\description{
Summarize basic model properties from a lavaan/blavaan fit
}
\examples{
# info <- model_info(fit)
}
